Move roomType from entities.icons to entities.
