+++ src/main/java/entities/RoomType.java
@@ -1,4 +1,4 @@
-package entities.icons;
+package entities;
 
 import javafx.scene.image.Image;
